Regret factor in ForwardBackward (#91)
This PR adds the possibility to specify a regret factor to increase the
stepsize gamma at every iteration, if adaptive, before backtracking.
Recent results provide convergence guarantees even without a maximum
value on gamma [Theorem 14,
[arXiv:2208.00779v2](https://arxiv.org/abs/2208.00799)].

This feature was implemented for ForwardBackward only. Although it does
not seem a good fit for accelerated methods like ZeroFPR, PANOC, and
PANOCplus, at least when based on quasi-Newton directions, it is unclear
whether the FastForwardBackward solver could benefit from it. See
discussion in #86 .

Practical performance seems to improve with values regret_gamma close to
1. Tests and references have also been updated accordingly.